Possibility to navigate to SD Card-Root-usr/share and file adding in FileManagerPage possible

+++ tide/qml/python/addFile.py
@@ -8,6 +8,10 @@ def createFile(fileName,ext,path):
         return "File already exists."
     else:
         try:
-            copyfile("/usr/share/harbour-tide/qml/templates/template"+ext, path +"/"+ fileName + ext)
+            if os.path.exists("/usr/share/harbour-tide/qml/templates/template"+ext):
+                copyfile("/usr/share/harbour-tide/qml/templates/template"+ext, path +"/"+ fileName + ext)
+            else:
+                file = open(path +"/"+ fileName + ext, 'a+', encoding = "utf-8")
+                file.close()
         except:
             raise
